Move asset class to model package.
